Move method to test file
